Move generic test functions, like register_dataset, to their own file
